+++ engine/mainui/controls/Field.h
@@ -47,7 +47,7 @@ public:
 	void SetBuffer( const char *buffer )
 	{
 		Q_strncpy( szBuffer, buffer, UI_MAX_FIELD_LINE );
-		iCursor = strlen( szBuffer );
+		iCursor = (int)strlen( szBuffer );
 		iScroll = g_FontMgr.CutText( font, szBuffer, m_scChSize, iRealWidth, true );
 	}
